splitted actions panel component from data collector
