+++ src/DataCollector/component.tsx
@@ -1,8 +1,6 @@
 import React from 'react';
-import { Container, Grid, TextField, Button } from '@material-ui/core';
-import { Autocomplete } from '@material-ui/lab';
-import { makeStyles } from '@material-ui/core/styles';
-import _ from 'lodash';
+import { Container } from '@material-ui/core';
+import ActionsPanel from './components/ActionsPanel';
 import Cameras from './components/Cameras';
 import useDataCollectorEffects from './hooks';
 
@@ -17,16 +15,6 @@ type Props = {
   createGroundTruthIsLoading: boolean;
 };
 
-const useStyles = makeStyles({
-  button: {
-    backgroundColor: '#2ca0f7',
-    color: 'white',
-    height: '55px',
-    textTransform: 'none',
-    fontFamily: 'Exo 2',
-  },
-});
-
 const DataCollector: React.FC<Props> = ({
   items,
   makePrediction,
@@ -37,7 +25,6 @@ const DataCollector: React.FC<Props> = ({
   createGroundTruthHasError,
   createGroundTruthIsLoading,
 }) => {
-  const classes = useStyles();
   useDataCollectorEffects(
     makePrediction,
     addItemsToReciept,
@@ -46,47 +33,18 @@ const DataCollector: React.FC<Props> = ({
     gt,
     predictions
   );
-  const onCaptureAll = (event: React.MouseEvent<HTMLButtonElement>) => {
-    event.preventDefault();
-    makePrediction();
-  };
   return (
     <Container style={{ marginTop: '50px' }} maxWidth="md">
-      <Grid container spacing={4} justify="center">
-        <Grid item>
-          <Autocomplete
-            id="labels"
-            onChange={(event, value) => setGroundTruth(value)}
-            options={_.uniqBy(Object.values(items), 'label')}
-            getOptionLabel={(option) => option.label || ''}
-            style={{ width: 300 }}
-            renderInput={(params) => (
-              <TextField {...params} label="Ground Truth" variant="outlined" />
-            )}
-          />
-        </Grid>
-        <Grid item>
-          <Button
-            variant="contained"
-            disableElevation
-            className={classes.button}
-            onClick={onCaptureAll}
-          >
-            Capture All (shift)
-          </Button>
-        </Grid>
-        <Grid item>
-          <Button
-            variant="contained"
-            disableElevation
-            className={classes.button}
-            disabled={createGroundTruthIsLoading || createGroundTruthHasError}
-            onClick={() => addItemsToReciept(gt, predictions)}
-          >
-            Save All (space)
-          </Button>
-        </Grid>
-      </Grid>
+      <ActionsPanel
+        items={items}
+        predictions={predictions}
+        makePrediction={makePrediction}
+        setGroundTruth={setGroundTruth}
+        addItemsToReciept={addItemsToReciept}
+        gt={gt}
+        createGroundTruthHasError={createGroundTruthHasError}
+        createGroundTruthIsLoading={createGroundTruthIsLoading}
+      />
       <Cameras
         predictions={predictions}
         makePrediction={makePrediction}
